Add DoSqliteDriver, typegres/core + do-sqlite entries, codegen fixes
