+++ src/database.test.ts
@@ -2,8 +2,8 @@ import { test, expect, beforeAll, afterAll } from "vitest";
 import { Int8, Text } from "./types/postgres";
 import { sql } from "./builder/sql";
 import { setupDb, db, conn } from "./test-helpers";
-import { PgDriver } from "./driver";
-import type { Driver } from "./driver";
+import { PgDriver } from "./drivers/pg";
+import type { Driver } from "./drivers/types";
 import { requireDatabaseUrl } from "./pg";
 import type { Connection } from "./database";
 import { Database } from "./database";
