+++ src/tables/generate.ts
@@ -242,7 +242,8 @@ const newFile = (
 
   // Split imports: dialect-specific type classes come from
   // `typeImportPath`; dialect-agnostic runtime helpers (`expose`,
-  // `sql`) come from the top `typegres` barrel. Symbols within each
+  // `sql`) come from `typegres` (no optional node driver peers —
+  // safe for Workers / Durable Object bundles). Symbols within each
   // group sorted for stable diffs.
   const runtimeSyms = ["expose", ...(hasDefault ? ["sql"] : [])].sort();
   const typeSyms = [...typeClasses].sort();
