+++ tsdown.config.ts
@@ -20,10 +20,10 @@ export default defineConfig([
   // loaders — those use `require` / `__filename` and would trip Node's
   // CJS/ESM mixed-mode check when a consumer runs the CLI.
   {
-    entry: ["src/index.ts", "src/config.ts", "src/builder/sql.ts", "src/types/postgres/index.ts", "src/types/sqlite/index.ts", "src/cli.ts", "src/exoeval/index.ts"],
+    entry: ["src/index.ts", "src/config.ts", "src/builder/sql.ts", "src/types/postgres/index.ts", "src/types/sqlite/index.ts", "src/cli.ts", "src/exoeval/index.ts", "src/capnweb/shim.ts", "src/drivers/do.ts", "src/drivers/pg.ts", "src/drivers/pglite.ts", "src/drivers/sqlite.ts"],
     format: ["esm"],
     clean: true,
-    deps: { neverBundle: ["pg", "@electric-sql/pglite", "better-sqlite3"] },
+    deps: { neverBundle: ["pg", "@electric-sql/pglite", "better-sqlite3", "capnweb"] },
     plugins: [swcPlugin()],
   },
   // Playground single-file bundle for the site's Monaco + esbuild-wasm
